RavenDB-26630 Use the same component for channels
