+++ src/Raven.AiAppliance.Web/src/pages/apps/app-channels.tsx
@@ -1,0 +1,8 @@
+import { useParams } from "react-router";
+import { ChannelsSection } from "@/pages/apps/channels-section";
+
+export function AppChannels() {
+    const { slug = "" } = useParams();
+
+    return <ChannelsSection slug={slug} />;
+}
